DBZ-8340 Support compressed binlog events in event listeners
CountDownEventListener is used in many tests for waiting for events in
the DB. As it's based on the event type, it doesn't work for compressed
binlog, when events are `TRANSACTION_PAYLOAD` events.
To be able to run the tests also with compressed binlog without any
major rewrite, support `TRANSACTION_PAYLOAD` data in this helper class
and process all events contained in `TRANSACTION_PAYLOAD` event.

Similar to `CapturingEventListener`.